@@ -3,10 +3,11 @@ author: Dart Team <misc@dartlang.org>
 description: The UI client for a web based interactive Dart service.
 homepage: https://github.com/dart-lang/dart-pad
 environment:
-  sdk: ">=2.10.4 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   codemirror: ^0.6.0-0
+  collection: ^1.15.0
   fluttering_phrases: ^0.3.1
   http: ^0.13.0
   json_annotation: ^4.0.0
@@ -28,7 +29,6 @@ dev_dependencies:
   build_runner: ^2.0.0
   build_test: ^2.0.0
   build_web_compilers: ^3.0.0
-  collection: ^1.14.10
   git: ^2.0.0
   grinder: ^0.9.0
   json_serializable: ^4.0.0
@@ -36,3 +36,10 @@ dev_dependencies:
   shelf: ^1.0.0
   shelf_static: ^1.0.0
   test: ^1.15.7
+
+# waiting for the next material-components-web release.
+# Once released, it will need to be rolled into package:mdc_web.
+#
+# https://github.com/material-components/material-components-web/pull/7158
+dependency_overrides:
+  sass: 1.32.10
