@@ -8,10 +8,10 @@ abstract class WorkshopFetcher {
   Future<Workshop> fetch();
 
   factory WorkshopFetcher.github({
-    @required String owner,
-    @required String repo,
-    String ref,
-    String path,
+    required String owner,
+    required String repo,
+    String? ref,
+    String? path,
   }) =>
       GithubWorkshopFetcher(owner: owner, repo: repo, path: path, ref: ref);
 
