@@ -12,19 +12,19 @@ abstract class WorkshopFetcherImpl implements WorkshopFetcher {
   Future<Workshop> fetch() async {
     var metadata = await fetchMeta();
     var steps = await fetchSteps(metadata);
-    return Workshop(metadata.name, metadata.type, steps);
+    return Workshop(metadata.name!, metadata.type!, steps);
   }
 
   Future<Meta> fetchMeta() async {
     var contents = await loadFileContents(['meta.yaml']);
-    return checkedYamlDecode(contents, (Map m) => Meta.fromJson(m));
+    return checkedYamlDecode(contents, (Map? m) => Meta.fromJson(m!));
   }
 
   Future<List<Step>> fetchSteps(Meta metadata) async {
     // Fetch each step in parallel and place the results in the original order.
     var futures = <Future<Step>>[];
-    for (var i = 0; i < metadata.steps.length; i++) {
-      var config = metadata.steps[i];
+    for (var i = 0; i < metadata.steps!.length; i++) {
+      var config = metadata.steps![i];
       futures.add(fetchStep(config));
     }
     return Future.wait(futures);
@@ -32,9 +32,9 @@ abstract class WorkshopFetcherImpl implements WorkshopFetcher {
 
   Future<Step> fetchStep(StepConfiguration config) async {
     var directory = config.directory;
-    String instructions;
-    String snippet;
-    String /*?*/ solution;
+    late String instructions;
+    late String snippet;
+    String? solution;
 
     var futures = <Future<String>>[];
 
