@@ -2,7 +2,7 @@ class Step {
   final String name;
   final String instructions;
   final String snippet;
-  final String solution;
+  final String? solution;
 
   Step(this.name, this.instructions, this.snippet, {this.solution});
 }
