@@ -2,6 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'package:collection/collection.dart' show IterableExtension;
 import 'package:dart_pad/elements/elements.dart';
 import 'package:mdc_web/mdc_web.dart';
 
@@ -11,7 +12,7 @@ class MaterialTabController extends TabController {
   MaterialTabController(this.tabBar);
 
   @override
-  Future selectTab(String tabName) async {
+  Future selectTab(String? tabName) async {
     var tab = tabs.firstWhere((t) => t.name == tabName);
     var idx = tabs.indexOf(tab);
 
@@ -26,7 +27,7 @@ class MaterialTabController extends TabController {
 
   void setTabVisibility(String tabName, bool visible) {
     tabs
-        .firstWhere((t) => t.name == tabName, orElse: () => null)
+        .firstWhereOrNull((t) => t.name == tabName)
         ?.toggleAttr('hidden', !visible);
   }
 }
