@@ -13,19 +13,19 @@ void main() {
     setUp(() => embed.init(embed.EmbedOptions(embed.EmbedMode.flutter)));
 
     test('Editor tab is selected at init', () {
-      final editorTab = querySelector('#editor-tab');
+      final editorTab = querySelector('#editor-tab')!;
       expect(editorTab.attributes.keys, contains('aria-selected'));
       expect(editorTab.classes, contains('mdc-tab--active'));
     });
 
     test('Test tab is not selected at init', () {
-      final testTab = querySelector('#test-tab');
+      final testTab = querySelector('#test-tab')!;
       expect(testTab.attributes.keys, isNot(contains('aria-selected')));
       expect(testTab.classes, isNot(contains('mdc-tab--active')));
     });
 
     test('Test tab is selected when clicked.', () {
-      final testTab = querySelector('#test-tab');
+      final testTab = querySelector('#test-tab')!;
       testTab.dispatchEvent(Event('click'));
 
       expect(testTab.attributes.keys, contains('aria-selected'));
@@ -33,10 +33,10 @@ void main() {
     });
 
     test('Editor tab is not selected when test tab is clicked', () {
-      final testTab = querySelector('#test-tab');
+      final testTab = querySelector('#test-tab')!;
       testTab.dispatchEvent(Event('click'));
 
-      final editorTab = querySelector('#editor-tab');
+      final editorTab = querySelector('#editor-tab')!;
       expect(editorTab.attributes.keys, isNot(contains('aria-selected')));
       expect(editorTab.classes, isNot(contains('mdc-tab--active')));
     });
