@@ -169,14 +169,14 @@ build() {
 
 /// Formats a map of argument key and values to be passed as `dart2js_args` for
 /// webdev.
-String _formatDart2jsArgs(Map<String, String> args) {
+String _formatDart2jsArgs(Map<String, String?> args) {
   var values = args.entries.map((entry) => '"-D${entry.key}=${entry.value}"');
   return '[${values.join(',')}]';
 }
 
 /// Formats a map of argument key and values to be passed as DDC environment
 /// variables.
-String _formatDdcArgs(Map<String, String> args) {
+String _formatDdcArgs(Map<String, String?> args) {
   var values = args.entries.map((entry) => '"${entry.key}":"${entry.value}"');
   return '{${values.join(',')}}';
 }
@@ -192,7 +192,7 @@ coverage() {
   coveralls.run([
     'report',
     '--token',
-    _env['COVERAGE_TOKEN'],
+    _env['COVERAGE_TOKEN']!,
     '--retry',
     '2',
     '--exclude-test-files',
@@ -285,7 +285,7 @@ class ConstTaskArgs implements TaskArgs {
   bool hasOption(String name) => _options.containsKey(name);
 
   @override
-  String getOption(String name) => _options[name];
+  String? getOption(String name) => _options[name];
 
   @override
   List<String> get arguments => throw UnimplementedError();
